requirements.txt
streamlit==1.28.0
plotly==5.18.0
pandas==2.0.3
gigachat==0.2.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
python-docx==1.1.0
pdfplumber==0.10.3
speechrecognition==3.10.0
pydub==0.25.1
pyaudio==0.2.11
vosk==0.3.45
sounddevice==0.4.6
pyttsx3==2.90